Initial implementation of MEMORY_STRATEGY
I chose to use a single buffer for all modules and levels, assuming that the user
would already fine-tune the memory queue using the per-level strategy flags.
Could be changed to something similar to Logfile if there is a need.

My implementation of a dynamic (sic) circular buffer is a bit messy and could likely
be improved a lot... Especially the algorithm to resize the buffer :D

Closes #1.